Move Fragment Tests to `test_fragments.py`
